+++ contrib/platform/src/com/sun/jna/platform/win32/FlagEnum.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2014 Martin Steiger
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.sun.jna.platform.win32;
+
+/**
+ * An interface for enum that can be combined to a 
+ * set based on an integer value 
+ * @author Martin Steiger
+ */
+public interface FlagEnum
+{
+	/**
+	 * @return the flag value - usually 2^n
+	 */
+	int getFlag();
+}